Move HedvigApp outside of the activity and to another file
