+++ app/app/src/main/kotlin/com/hedvig/android/app/ui/HedvigApp.kt
@@ -1,131 +1,206 @@
 package com.hedvig.android.app.ui
 
-import androidx.compose.animation.core.AnimationVector4D
-import androidx.compose.animation.core.TwoWayConverter
-import androidx.compose.animation.core.animateValueAsState
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.WindowInsetsSides
-import androidx.compose.foundation.layout.asPaddingValues
-import androidx.compose.foundation.layout.consumeWindowInsets
-import androidx.compose.foundation.layout.displayCutout
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.only
-import androidx.compose.foundation.layout.systemBars
-import androidx.compose.foundation.layout.union
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
+import android.graphics.Color
+import androidx.activity.SystemBarStyle
+import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.unit.LayoutDirection
-import androidx.compose.ui.unit.dp
+import androidx.compose.runtime.produceState
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalLifecycleOwner
+import androidx.compose.ui.platform.LocalUriHandler
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.lifecycle.repeatOnLifecycle
+import androidx.navigation.NavBackStackEntry
+import androidx.navigation.NavDestination.Companion.hierarchy
+import androidx.navigation.NavHostController
 import coil.ImageLoader
-import com.hedvig.android.app.navigation.HedvigNavHost
+import com.hedvig.android.auth.AuthStatus
+import com.hedvig.android.auth.AuthTokenService
+import com.hedvig.android.core.appreview.WaitUntilAppReviewDialogShouldBeOpenedUseCase
 import com.hedvig.android.core.buildconstants.HedvigBuildConstants
-import com.hedvig.android.core.designsystem.material3.motion.MotionTokens
+import com.hedvig.android.core.demomode.DemoManager
+import com.hedvig.android.core.demomode.Provider
+import com.hedvig.android.core.designsystem.theme.HedvigTheme
+import com.hedvig.android.data.paying.member.GetOnlyHasNonPayingContractsUseCase
+import com.hedvig.android.data.settings.datastore.SettingsDataStore
+import com.hedvig.android.feature.force.upgrade.ForceUpgradeBlockingScreen
+import com.hedvig.android.feature.login.navigation.LoginDestination
+import com.hedvig.android.featureflags.FeatureManager
 import com.hedvig.android.language.LanguageService
-import com.hedvig.android.market.Market
+import com.hedvig.android.logger.logcat
+import com.hedvig.android.market.MarketManager
 import com.hedvig.android.navigation.activity.ActivityNavigator
 import com.hedvig.android.navigation.core.HedvigDeepLinkContainer
+import com.hedvig.android.notification.badge.data.tab.TabNotificationBadgeService
+import com.kiwi.navigationcompose.typed.createRoutePattern
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.receiveAsFlow
 
 @Composable
 internal fun HedvigApp(
-  hedvigAppState: HedvigAppState,
-  hedvigDeepLinkContainer: HedvigDeepLinkContainer,
+  navHostController: NavHostController,
+  tabNotificationBadgeService: TabNotificationBadgeService,
+  settingsDataStore: SettingsDataStore,
+  getOnlyHasNonPayingContractsUseCase: Provider<GetOnlyHasNonPayingContractsUseCase>,
+  featureManager: FeatureManager,
+  splashIsRemovedSignal: Channel<Unit>,
   activityNavigator: ActivityNavigator,
-  finishApp: () -> Unit,
-  shouldShowRequestPermissionRationale: (String) -> Boolean,
-  openUrl: (String) -> Unit,
-  onOpenEmailApp: () -> Unit,
-  market: Market,
+  authTokenService: AuthTokenService,
+  demoManager: DemoManager,
+  hedvigDeepLinkContainer: HedvigDeepLinkContainer,
+  marketManager: MarketManager,
   imageLoader: ImageLoader,
   languageService: LanguageService,
   hedvigBuildConstants: HedvigBuildConstants,
+  waitUntilAppReviewDialogShouldBeOpenedUseCase: WaitUntilAppReviewDialogShouldBeOpenedUseCase,
+  enableEdgeToEdge: (SystemBarStyle) -> Unit,
+  shouldShowRequestPermissionRationale: (String) -> Boolean,
+  goToPlayStore: () -> Unit,
+  openEmailApp: () -> Unit,
+  finishApp: () -> Unit,
+  tryShowAppStoreReviewDialog: () -> Unit,
+  windowSizeClass: WindowSizeClass,
 ) {
-  Surface(
-    color = MaterialTheme.colorScheme.background,
-    contentColor = MaterialTheme.colorScheme.onBackground,
-    modifier = Modifier.fillMaxSize(),
-  ) {
-    NavigationSuite(
-      navigationSuiteType = hedvigAppState.navigationSuiteType,
-      topLevelGraphs = hedvigAppState.topLevelGraphs.collectAsState().value,
-      topLevelGraphsWithNotifications = hedvigAppState.topLevelGraphsWithNotifications.collectAsState().value,
-      currentDestination = hedvigAppState.currentDestination,
-      onNavigateToTopLevelGraph = hedvigAppState::navigateToTopLevelGraph,
-    ) {
-      HedvigNavHost(
-        hedvigAppState = hedvigAppState,
-        hedvigDeepLinkContainer = hedvigDeepLinkContainer,
-        activityNavigator = activityNavigator,
-        finishApp = finishApp,
-        shouldShowRequestPermissionRationale = shouldShowRequestPermissionRationale,
-        openUrl = openUrl,
-        onOpenEmailApp = onOpenEmailApp,
-        imageLoader = imageLoader,
-        market = market,
-        languageService = languageService,
-        hedvigBuildConstants = hedvigBuildConstants,
-        modifier = Modifier
-          .fillMaxHeight()
-          .weight(1f)
-          .animatedNavigationBarInsetsConsumption(hedvigAppState),
+  val hedvigAppState = rememberHedvigAppState(
+    windowSizeClass = windowSizeClass,
+    tabNotificationBadgeService = tabNotificationBadgeService,
+    settingsDataStore = settingsDataStore,
+    getOnlyHasNonPayingContractsUseCase = getOnlyHasNonPayingContractsUseCase,
+    featureManager = featureManager,
+    navHostController = navHostController,
+  )
+  val darkTheme = hedvigAppState.darkTheme
+  HedvigTheme(darkTheme = darkTheme) {
+    EnableEdgeToEdgeSideEffect(darkTheme, splashIsRemovedSignal, enableEdgeToEdge)
+    val mustForceUpdate by hedvigAppState.mustForceUpdate.collectAsStateWithLifecycle()
+    if (mustForceUpdate) {
+      ForceUpgradeBlockingScreen(
+        goToPlayStore = goToPlayStore,
       )
+    } else {
+      TryShowAppStoreReviewDialogEffect(
+        authTokenService,
+        waitUntilAppReviewDialogShouldBeOpenedUseCase,
+        tryShowAppStoreReviewDialog,
+      )
+      LogoutOnInvalidCredentialsEffect(hedvigAppState, authTokenService, demoManager)
+      val deepLinkFirstUriHandler = DeepLinkFirstUriHandler(
+        navController = hedvigAppState.navController,
+        delegate = SafeAndroidUriHandler(LocalContext.current),
+      )
+      CompositionLocalProvider(LocalUriHandler provides deepLinkFirstUriHandler) {
+        HedvigAppUi(
+          hedvigAppState = hedvigAppState,
+          hedvigDeepLinkContainer = hedvigDeepLinkContainer,
+          activityNavigator = activityNavigator,
+          shouldShowRequestPermissionRationale = shouldShowRequestPermissionRationale,
+          openUrl = deepLinkFirstUriHandler::openUri,
+          onOpenEmailApp = openEmailApp,
+          finishApp = finishApp,
+          market = marketManager.market.collectAsStateWithLifecycle().value,
+          imageLoader = imageLoader,
+          languageService = languageService,
+          hedvigBuildConstants = hedvigBuildConstants,
+        )
+      }
+    }
+  }
+}
+
+@Composable
+private fun EnableEdgeToEdgeSideEffect(
+  darkTheme: Boolean,
+  splashIsRemovedSignal: Channel<Unit>,
+  enableEdgeToEdge: (SystemBarStyle) -> Unit,
+) {
+  val splashIsRemovedIndex by produceState(0) {
+    splashIsRemovedSignal.receiveAsFlow().collectLatest { value = value + 1 }
+  }
+  DisposableEffect(darkTheme, splashIsRemovedIndex) {
+    enableEdgeToEdge(
+      when (darkTheme) {
+        true -> SystemBarStyle.dark(Color.TRANSPARENT)
+        false -> SystemBarStyle.light(Color.TRANSPARENT, Color.TRANSPARENT)
+      },
+    )
+    onDispose {}
+  }
+}
+
+@Composable
+private fun TryShowAppStoreReviewDialogEffect(
+  authTokenService: AuthTokenService,
+  waitUntilAppReviewDialogShouldBeOpenedUseCase: WaitUntilAppReviewDialogShouldBeOpenedUseCase,
+  tryShowAppStoreReviewDialog: () -> Unit,
+) {
+  val REVIEW_DIALOG_DELAY_MILLIS = 2000L
+  val lifecycle = LocalLifecycleOwner.current.lifecycle
+  LaunchedEffect(lifecycle) {
+    lifecycle.repeatOnLifecycle(Lifecycle.State.RESUMED) {
+      authTokenService.authStatus.first { it is AuthStatus.LoggedIn }
+      waitUntilAppReviewDialogShouldBeOpenedUseCase.invoke()
+      delay(REVIEW_DIALOG_DELAY_MILLIS)
+      tryShowAppStoreReviewDialog()
     }
   }
 }
 
 /**
- * Animates how we consume the insets, so that when we leave a screen which does not consume any insets, and we enter a
- * screen which does (by showing the bottom nav for example) then we don't want the outgoing screen to have its
- * contents snap to the bounds of the new insets immediately. This animation makes this visual effect look much more
- * fluid.
+ * Automatically logs out when we are no longer in demo mode and we are also not considered to have active tokens
  */
-@OptIn(ExperimentalLayoutApi::class)
-private fun Modifier.animatedNavigationBarInsetsConsumption(hedvigAppState: HedvigAppState) = composed {
-  val density = LocalDensity.current
-  val insetsToConsume = when (hedvigAppState.navigationSuiteType) {
-    NavigationSuiteType.NavigationBar -> WindowInsets.systemBars.only(WindowInsetsSides.Bottom).asPaddingValues(density)
-    NavigationSuiteType.NavigationRail -> WindowInsets.systemBars.union(WindowInsets.displayCutout)
-      .only(WindowInsetsSides.Left).asPaddingValues(density)
-
-    else -> PaddingValues(0.dp)
+@Composable
+private fun LogoutOnInvalidCredentialsEffect(
+  hedvigAppState: HedvigAppState,
+  authTokenService: AuthTokenService,
+  demoManager: DemoManager,
+) {
+  val authStatusLog: (AuthStatus?) -> Unit = { authStatus ->
+    logcat {
+      buildString {
+        append("Owner: LoggedInActivity | Received authStatus: ")
+        append(
+          when (authStatus) {
+            is AuthStatus.LoggedIn -> "LoggedIn"
+            AuthStatus.LoggedOut -> "LoggedOut"
+            null -> "null"
+          },
+        )
+      }
+    }
   }
-
-  val paddingValuesVectorConverter: TwoWayConverter<PaddingValues, AnimationVector4D> = TwoWayConverter(
-    convertToVector = { paddingValues ->
-      AnimationVector4D(
-        paddingValues.calculateLeftPadding(LayoutDirection.Ltr).value,
-        paddingValues.calculateRightPadding(LayoutDirection.Ltr).value,
-        paddingValues.calculateTopPadding().value,
-        paddingValues.calculateBottomPadding().value,
-      )
-    },
-    convertFromVector = { animationVector4d ->
-      val leftPadding = animationVector4d.v1
-      val rightPadding = animationVector4d.v2
-      val topPadding = animationVector4d.v3
-      val bottomPadding = animationVector4d.v4
-      PaddingValues(
-        start = leftPadding.dp,
-        end = rightPadding.dp,
-        top = topPadding.dp,
-        bottom = bottomPadding.dp,
-      )
-    },
-  )
-  val animatedInsetsToConsume: PaddingValues by animateValueAsState(
-    targetValue = insetsToConsume,
-    typeConverter = paddingValuesVectorConverter,
-    animationSpec = tween(MotionTokens.DurationMedium1.toInt()),
-    label = "Padding values inset animation",
-  )
-  consumeWindowInsets(animatedInsetsToConsume)
+  val lifecycle = LocalLifecycleOwner.current.lifecycle
+  LaunchedEffect(lifecycle, hedvigAppState, authTokenService, demoManager) {
+    val loginGraphRoute = createRoutePattern<LoginDestination>()
+    lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+      combine(
+        authTokenService.authStatus.onEach(authStatusLog).filterNotNull().distinctUntilChanged(),
+        demoManager.isDemoMode().distinctUntilChanged(),
+      ) { authStatus: AuthStatus, isDemoMode: Boolean ->
+        authStatus to isDemoMode
+      }.collect { (authStatus, isDemoMode) ->
+        val navBackStackEntry: NavBackStackEntry = hedvigAppState.navController.currentBackStackEntryFlow.first()
+        val isLoggedOut = navBackStackEntry.destination.hierarchy.any { navDestination ->
+          navDestination.route?.contains(loginGraphRoute) == true
+        }
+        if (isLoggedOut) {
+          return@collect
+        }
+        if (!isDemoMode && authStatus !is AuthStatus.LoggedIn) {
+          hedvigAppState.navigateToLoggedOut()
+        }
+      }
+    }
+  }
 }
